backend/requirements.txt
fastapi[standard]
passlib
sqlalchemy
pydantic[email]
asyncpg
python-jose
alembic
black
pytest
bcrypt
uvicorn
psycopg2-binary
pydantic_settings
pytest-mock